windows support

+++ setup.py
@@ -8,27 +8,29 @@ from setuptools import setup
     we exec cmake from cmd here.
 '''
 
+is_win = os.name=='nt'
+
 
 class MyInstall(install):
     def run(self):
-        if not os.path.exists('multicore_tsne/release'):
-            os.makedirs('multicore_tsne/release')
-        else:
-            os.system('rm -rf multicore_tsne/release/')
-            os.makedirs('multicore_tsne/release')
+        if not is_win:
+            if not os.path.exists('multicore_tsne/release'):
+                os.makedirs('multicore_tsne/release')
+            else:
+                os.system('rm -rf multicore_tsne/release/')
+                os.makedirs('multicore_tsne/release')
 
-        os.chdir('multicore_tsne/release/')
-        return_val = os.system('cmake -DCMAKE_BUILD_TYPE=RELEASE ..')
+            os.chdir('multicore_tsne/release/')
+            return_val = os.system('cmake -DCMAKE_BUILD_TYPE=RELEASE ..')
 
-        if return_val != 0:
-            print('cannot find cmake')
-            exit(-1)
+            if return_val != 0:
+                print('cannot find cmake')
+                exit(-1)
 
-        os.system('make VERBOSE=1')
-        os.chdir('../..')
-        print(os.getcwd())
-        os.system(
-            'cp multicore_tsne/release/libtsne_multicore.so python/libtsne_multicore.so')
+            os.system('make VERBOSE=1')
+            os.chdir('../..')
+            print(os.getcwd())
+            os.system('cp multicore_tsne/release/tsne_multicore.dll python/libtsne_multicore.dll')
         install.run(self)
 
 
@@ -47,7 +49,7 @@ setup(
 
     packages=['MulticoreTSNE'],
     package_dir={'MulticoreTSNE': 'python'},
-    package_data={'MulticoreTSNE': ['multicore_tsne.so']},
+    package_data={'MulticoreTSNE': ['libtsne_multicore.' + ('dll' if is_win else 'so')]},
     include_package_data=True,
 
     cmdclass={"install": MyInstall},
